src/ie/gmit/sw/test/ParseTest.java
package ie.gmit.sw.test;

import org.junit.*;
import static org.junit.Assert.*;

import java.util.ArrayList;
import java.util.HashMap;

import ie.gmit.sw.wordcloud.*;

public class ParseTest {

	private Parse parseOne = null;
	private Parse parseTwo = null;
	
	@Before
	public void setup() throws Exception {
		parseOne = new ParseFile("stopwords.txt", "fileparse");
		parseTwo = new ParseURL("stopwords.txt", "urlparse");
	}
	
	@After
	public void tearDown(){
		parseOne = null;
		parseTwo = null;
	}
	
	@Test
	public void validStopWordsList() throws Exception {
		assertNotNull(parseOne.getStopWords());
		assertNotNull(parseTwo.getStopWords());
		parseOne.setStopWords(new ArrayList<String>());
		parseTwo.setStopWords(new ArrayList<String>());
		assertNotNull(parseOne.getStopWords());
		assertNotNull(parseTwo.getStopWords());
	}
	
	@Test
	public void validWordFrequencyMap() throws Exception {
		assertNotNull(parseOne.getWordFrequencyMap());
		assertNotNull(parseOne.getWordFrequencyMap());
		parseOne.setWordFrequencyMap(new HashMap<String, Integer>());
		parseTwo.setWordFrequencyMap(new HashMap<String, Integer>());
		assertNotNull(parseOne.getWordFrequencyMap());
		assertNotNull(parseOne.getWordFrequencyMap());
	}
	
	@Test(expected=Exception.class)
	public void nullWordFrequencyMap() throws Exception {
		parseOne.setWordFrequencyMap(null);
	}
}